Updated the way configs get the components for a current run. Now the components are read from a text file, so we only have to modify 1 file instead of several. Also updated some transition models to point to the correct MCS or PCS, and cleaned up some things in modules

+++ minos/data_generation/generate_repl_pop.py
@@ -229,8 +229,8 @@ def generate_replenishing(projections, scotland_mode, cross_validation, inflated
     final_repl['S7_physical_health'] = final_repl['S7_physical_health'].astype(int)
     final_repl['nutrition_quality_diff'] = final_repl['nutrition_quality_diff'].astype(int)
     final_repl['neighbourhood_safety'] = final_repl['neighbourhood_safety'].astype(int)
-    final_repl['chron_disease'] = final_repl['chron_disease'].astype(float)
-    final_repl['matdep'] = final_repl['matdep'].astype(float)
+    final_repl['chron_disease'] = final_repl['chron_disease'].astype(int)
+    final_repl['matdep'] = final_repl['matdep'].astype(int)
 
     US_utils.check_output_dir(output_dir)
     final_repl.to_csv(f'{output_dir}/replenishing_pop_2019-2070.csv', index=False)
