+++ minos/minosPipeline/RunPipeline.py
@@ -234,8 +234,8 @@ def type_check(data):
     data['S7_physical_health'] = data['S7_physical_health'].astype(int)
     data['nutrition_quality_diff'] = data['nutrition_quality_diff'].astype(int)
     data['neighbourhood_safety'] = data['neighbourhood_safety'].astype(int)
-    data['chron_disease'] = data['chron_disease'].astype(float)
-    data['matdep'] = data['matdep'].astype(float)
+    data['chron_disease'] = data['chron_disease'].astype(int)
+    data['matdep'] = data['matdep'].astype(int)
 
     return data
